requirements.txt
langchain
tiktoken
matplotlib
seaborn
tqdm
beautifulsoup4
pinecone-client
numpy
python-dotenv
openai
langchain
pypdf
mysql-connector-python
